requirements.txt
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.1
tqdm==4.66.2
loguru==0.7.2
pydantic==1.10.13
python-dateutil==2.8.2
aiohttp==3.9.3
asyncio==3.4.3
PyPDF2==3.0.1
deep-translator==1.11.4
psycopg2-binary==2.9.9
sumy>=0.11.0
nltk>=3.8.1
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99 